Report merge commit to deploy times in datadog
This reports a new stat, `ci.merge_commit_to_deploy_time`. It goes off the commit time of the head commit, which is the time when something is added to a GH queue.

I've kept the tags pretty much the same as other dd metrics called via these actions, but I've also included the environment.

Weird bits:
- "Commit to deploy time" with no "merge" already has meaning as CTD time, so I couldnt use it here.
- This is not the most accurate as the job gets triggered a little while after the deploy happens, but this is the least brittle way of doing this as our deploy step code gets shuffled around every 2 weeks and passing outputs could easily break.